+++ lib/modules/ui/ui_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: ui
-// Updated at: 2026-07-20 23:10:27
-// Commit: perf(ui): apply RepaintBoundary on animated ui widgets
+// Updated at: 2026-07-21 12:51:26
+// Commit: feat(ui): integrate custom telemetry logger in ui
 
 class uiComponentState {
-  final String commitRef = '2969';
-  final String lastMessage = 'perf(ui): apply RepaintBoundary on animated ui widgets';
-  final DateTime timestamp = DateTime.parse('2026-07-20T23:10:27+05:00');
+  final String commitRef = '2989';
+  final String lastMessage = 'feat(ui): integrate custom telemetry logger in ui';
+  final DateTime timestamp = DateTime.parse('2026-07-21T12:51:26+05:00');
 }